+++ packages/toolbox-core/CHANGELOG.md
@@ -1,5 +1,47 @@
 # Changelog
 
+## [0.5.1](https://github.com/googleapis/mcp-toolbox-sdk-python/compare/toolbox-core-v0.5.0...toolbox-core-v0.5.1) (2025-09-17)
+
+
+### Bug Fixes
+
+* **toolbox-core:** Use typing.Annotated for reliable parameter descriptions instead of docstrings ([#371](https://github.com/googleapis/mcp-toolbox-sdk-python/issues/371)) ([eb76680](https://github.com/googleapis/mcp-toolbox-sdk-python/commit/eb76680d24c344241f3c15293cd12904bbf91b0d))
+
+
+### Documentation
+
+* Update langgraph sample in toolbox-core ([#366](https://github.com/googleapis/mcp-toolbox-sdk-python/issues/366)) ([fe35082](https://github.com/googleapis/mcp-toolbox-sdk-python/commit/fe35082104d8039986eccbdab937f5f3e8b2042b))
+
+
+### Miscellaneous Chores
+
+* Remove redundant test for test_add_auth_token_getter_unused_token ([#347](https://github.com/googleapis/mcp-toolbox-sdk-python/issues/347)) ([dccaf1b](https://github.com/googleapis/mcp-toolbox-sdk-python/commit/dccaf1bd70e4e788fd80a9ef952aede4549b2fc7))
+* Remove duplicate header check during initialization ([#357](https://github.com/googleapis/mcp-toolbox-sdk-python/issues/357)) ([888170b](https://github.com/googleapis/mcp-toolbox-sdk-python/commit/888170b3c34744089fb3b6f2b6f613c4cd718a89))
+
+
+## [0.5.0](https://github.com/googleapis/mcp-toolbox-sdk-python/compare/toolbox-core-v0.4.0...toolbox-core-v0.5.0) (2025-08-19)
+
+
+### ⚠ BREAKING CHANGES
+
+* **core:** Use ValueError for missing 'items' in array parameters ([#325](https://github.com/googleapis/mcp-toolbox-sdk-python/issues/325))
+
+### Features
+
+* **core:** Add support for map parameter type ([#324](https://github.com/googleapis/mcp-toolbox-sdk-python/issues/324)) ([6455ae3](https://github.com/googleapis/mcp-toolbox-sdk-python/commit/6455ae3b88774e7cd79f5c348e9f44bfc5424a18))
+
+
+### Bug Fixes
+
+* add optional clock_skew parameter to token helpers ([#333](https://github.com/googleapis/mcp-toolbox-sdk-python/issues/333)) ([aa0664f](https://github.com/googleapis/mcp-toolbox-sdk-python/commit/aa0664ff6d88d83bba4fc1a8ffe4766e4e8c8aa3))
+* **core:** Use ValueError for missing 'items' in array parameters ([#325](https://github.com/googleapis/mcp-toolbox-sdk-python/issues/325)) ([b43409e](https://github.com/googleapis/mcp-toolbox-sdk-python/commit/b43409ed220bbe89ab2521286239af203c067897))
+
+
+### Documentation
+
+* Enhance DEVELOPER.md for improved contribution guidelines ([#316](https://github.com/googleapis/mcp-toolbox-sdk-python/issues/316)) ([1d2be43](https://github.com/googleapis/mcp-toolbox-sdk-python/commit/1d2be4396d50d1119381149c7b1c4d1a778a0349))
+
+
 ## [0.4.0](https://github.com/googleapis/mcp-toolbox-sdk-python/compare/toolbox-core-v0.3.0...toolbox-core-v0.4.0) (2025-07-17)